requirements.txt
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
easydict>=1.10
